Create trial data frame from two files
This commit uses the two provided tsv files to create a dataframe that has a row for each patient observation that captures each of the three recorded values at that time.  It also provides patient characteristics alongside their observations and includes calculated fields for region and changes at the patient level between observations.

This commit also establishes a new naming convention for the variables where each variable leads with a lowercase unless it is an abbrevation such as ALT.