Longitudinal Analysis First Draft
Estiablishes the functions for the second tab of the application.  This tab allows for analyzing the change in the laboratory measurements over time.  The change in a specific laboratory measurement can be analyzed within groups.  Also tools to analyze the change in laboratory measurements within a single group are established.